==> 3_Preprocess.ipynb <==
### Script to create numerical encoding for categorical columns and create separate files for the encoded categories.

- Outputs:

    csv_imports/encoded

from numpy import array
from numpy import argmax
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import OneHotEncoder
import collections

import pandas as pd 

df = pd.read_csv('csv_imports/WN-edges.csv',index_col=[0])
print('Reading edges:\n')
print(df.head(1))
print()

# extract type codes (unique for every type).
TYPE_CAT = df[':TYPE'].unique()
print(TYPE_CAT)

df[':TYPE'] = df[':TYPE'].astype('category')
TYPE = df[':TYPE'].cat.codes.unique()
print('type code generated =>type_wn.csv')
print(TYPE)
print()

df1=pd.DataFrame(TYPE_CAT,columns=['type'])
df2=pd.DataFrame(TYPE,columns=['id'])
# print(df1)
# print(df2)

df_concat = pd.concat([df1,df2], axis=1)
df_concat.to_csv('csv_imports/encoded/type_wn.csv')

#encode the column type
# integer encode
label_encoder = LabelEncoder()
df[':TYPE'] = label_encoder.fit_transform(df[':TYPE'])
pd.DataFrame(df).to_csv('csv_imports/encoded/WN-edges-encoded.csv')
print('Numerical encoding completed: Edges(TYPE) => wn-edges-encoded.csv\n')
print(df.head(1))
print()

df = pd.read_csv('csv_imports/WN-nodes.csv',index_col=[0])
print('Reading nodes:\n')
print(df.head(1))
print()

# extract type codes (unique for every type).
TYPE_CAT = df['pos'].unique()
print(TYPE_CAT)

df['pos'] = df['pos'].astype('category')
TYPE = df['pos'].cat.codes.unique()
print('type code generated =>pos_wn.csv')
print(TYPE)
print()

df1=pd.DataFrame(TYPE_CAT,columns=['type'])
df2=pd.DataFrame(TYPE,columns=['id'])
# print(df1)
# print(df2)

df_concat = pd.concat([df1,df2], axis=1)
df_concat.to_csv('csv_imports/encoded/pos_wn.csv')

#encode the column type
# integer encode
label_encoder = LabelEncoder()
df['pos'] = label_encoder.fit_transform(df['pos'])
pd.DataFrame(df).to_csv('csv_imports/encoded/WN-nodes-encoded.csv')
print('Numerical encoding completed: Nodes(pos) => wn-nodes-encoded.csv\n')
print(df.head(1))
print()

df = pd.read_csv('csv_imports/encoded/WN-nodes-encoded.csv',index_col=[0])
print('Reading nodes:\n')
print(df.head(1))
print()

# extract type codes (unique for every type).
TYPE_CAT = df['subject'].unique()
print(TYPE_CAT)

df['subject'] = df['subject'].astype('category')
TYPE = df['subject'].cat.codes.unique()
print('type code generated =>subject_wn.csv')
print(TYPE)
print()

df1=pd.DataFrame(TYPE_CAT,columns=['type'])
df2=pd.DataFrame(TYPE,columns=['id'])
# print(df1)
# print(df2)

df_concat = pd.concat([df1,df2], axis=1)
df_concat.to_csv('csv_imports/encoded/subject_wn.csv')

#encode the column type
# integer encode
label_encoder = LabelEncoder()
df['subject'] = label_encoder.fit_transform(df['subject'])
pd.DataFrame(df).to_csv('csv_imports/encoded/WN-nodes-encoded.csv')
print('Numerical encoding completed: Nodes(Subject) => wn-nodes-encoded.csv\n')
print(df.head(1))
print()

df = pd.read_csv('csv_imports/synsets.csv',index_col=[0])
print('Reading nodes:\n')
print(df.head(1))
print()

# extract type codes (unique for every type).
TYPE_CAT = df['pos:string'].unique()
print(TYPE_CAT)

df['pos:string'] = df['pos:string'].astype('category')
TYPE = df['pos:string'].cat.codes.unique()
print('type code generated =>pos_synset.csv')
print(TYPE)
print()

df1=pd.DataFrame(TYPE_CAT,columns=['type'])
df2=pd.DataFrame(TYPE,columns=['id'])
print(df1)
print(df2)

df_concat = pd.concat([df1,df2], axis=1)
df_concat.to_csv('csv_imports/encoded/pos_synset.csv')


#encode the column type
# integer encode
label_encoder = LabelEncoder()
df['pos:string'] = label_encoder.fit_transform(df['pos:string'])
pd.DataFrame(df).to_csv('csv_imports/encoded/synsets-encoded.csv')
print('Numerical encoding completed: Synset(pos) => synsets-encoded.csv\n')
print(df.head(1))
print()

df = pd.read_csv('csv_imports/relationships.csv',index_col=[0])
print('Reading nodes:\n')
print(df.head(1))
print()

# extract type codes (unique for every type).
TYPE_CAT = df['dataset:string'].unique()
print(TYPE_CAT)

df['dataset:string'] = df['dataset:string'].astype('category')
TYPE = df['dataset:string'].cat.codes.unique()
print('type code generated =>dataset_rel.csv')
print(TYPE)
print()

df1=pd.DataFrame(TYPE_CAT,columns=['type'])
df2=pd.DataFrame(TYPE,columns=['id'])
print(df1)
print(df2)

df_concat = pd.concat([df1,df2], axis=1)
df_concat.to_csv('csv_imports/encoded/dataset_rel.csv')


#encode the column type
# integer encode
label_encoder = LabelEncoder()
df['dataset:string'] = label_encoder.fit_transform(df['dataset:string'])
pd.DataFrame(df).to_csv('csv_imports/encoded/relationships_encoded.csv')
print('Numerical encoding completed: Synset(pos) => relationships_encoded.csv\n')
print(df.head(1))
print()

df = pd.read_csv('csv_imports/encoded/relationships_encoded.csv',index_col=[0])
print('Reading nodes:\n')
print(df.head(1))
print()

# extract type codes (unique for every type).
TYPE_CAT = df[':TYPE'].unique()
print(TYPE_CAT)

df[':TYPE'] = df[':TYPE'].astype('category')
TYPE = df[':TYPE'].cat.codes.unique()
print('type code generated =>type_rel.csv')
print(TYPE)
print()

df1=pd.DataFrame(TYPE_CAT,columns=['type'])
df2=pd.DataFrame(TYPE,columns=['id'])
print(df1)
print(df2)

df_concat = pd.concat([df1,df2], axis=1)
df_concat.to_csv('csv_imports/encoded/type_rel.csv')


#encode the column type
# integer encode
label_encoder = LabelEncoder()
df[':TYPE'] = label_encoder.fit_transform(df[':TYPE'])
pd.DataFrame(df).to_csv('csv_imports/encoded/relationships_encoded.csv')
print('Numerical encoding completed: relationships(:TYPE) => relationships_encoded.csv\n')
print(df.head(1))
print()

df = pd.read_csv('csv_imports/words.csv',index_col=[0])
print('Reading nodes:\n')
print(df.head(1))
print()

# extract type codes (unique for every type).
TYPE_CAT = df['pos'].unique()
print(TYPE_CAT)

df['pos'] = df['pos'].astype('category')
TYPE = df['pos'].cat.codes.unique()
print('type code generated =>pos_words.csv')
print(TYPE)
print()

df1=pd.DataFrame(TYPE_CAT,columns=['type'])
df2=pd.DataFrame(TYPE,columns=['id'])
print(df1)
print(df2)

df_concat = pd.concat([df1,df2], axis=1)
df_concat.to_csv('csv_imports/encoded/pos_words.csv')


#encode the column type
# integer encode
label_encoder = LabelEncoder()
df['pos'] = label_encoder.fit_transform(df['pos'])
pd.DataFrame(df).to_csv('csv_imports/encoded/words_encoded.csv')
print('Numerical encoding completed: words(:TYPE) => words_encoded.csv\n')
print(df.head(1))
print()
