==> 4_LanguageModel_SingleEdge.ipynb <==
# Script to load wordnet, conceptnet data as separate graphs with single edge.

## Install pyTigerGraph

# Setup
!pip install pyTigerGraph

## Add Imports and Establish Initial Connection

# Imports
import pyTigerGraph as tg
import json
import pandas as pd

# Connection parameters
hostName = "https://language.i.tgcloud.io/"
userName = "tigergraph"
password = "tigergraph"

conn = tg.TigerGraphConnection(host=hostName, username=userName, password=password)

print("Connected")

## Peek into the data 

### ConceptNet - Word

# root='/content/drive/MyDrive/LanguageModel/Data_Processing/csv_imports/''
root='csv_imports/'
nodes = pd.read_csv(root+'words.csv')
nodes.head(1)

### ConceptNet - Synset

synsets = pd.read_csv(root+'synsets.csv')
synsets.head(1)

### ConceptNet - Edges

relationships = pd.read_csv(root+'relationships.csv',index_col=[0])
relationships.head(5)

### ConceptNet - Part of Speech

relationships = pd.read_csv(root+'encoded/pos_wn.csv',index_col=[0])
relationships.head()

##Clear graph if already solution has the same graph

# conn.gsql('''
# USE GLOBAL
# DROP ALL
# ''')

## Define and Publish the Schema

# DEFINE / CREATE ALL EDGES AND VERTICES 
results = conn.gsql('''
  USE GLOBAL
  CREATE VERTEX words(PRIMARY_ID id STRING, uri STRING, word STRING, pos STRING,definition STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
  CREATE DIRECTED EDGE type(FROM words, TO words, typeOf STRING) WITH REVERSE_EDGE="reverse_type"
  CREATE VERTEX word(PRIMARY_ID id STRING, name STRING, pos STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
  CREATE VERTEX synset(PRIMARY_ID id STRING, definition STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
  CREATE DIRECTED EDGE is_a(FROM synset, TO word|FROM word, TO synset, type STRING) WITH REVERSE_EDGE="reverse_is_a"
''')
print(results)

## WordNET

### Create WordNet Graph

results = conn.gsql('CREATE GRAPH WordNet(words,type,reverse_type)')
print(results)

conn.graphname="WordNet"
secret = conn.createSecret()
print(secret)
authToken = conn.getToken(secret)
authToken = authToken[0]
print(authToken)
# authToken = 'rc7reopbis1667ksgcppq5v5fb99p6s1'
conn = tg.TigerGraphConnection(host=hostName, graphname="WordNet", username=userName, password=password, apiToken=authToken)

def pprint(string):
  print(json.dumps(string, indent=2))

### Create Loading Jobs

#### Word

Let's take a look at what one of our files looks like so we can write a loading job.

- Here it's important to note that the `$0`, `$1` values line up with the columns of your data.
  In this example:
  - `$0` is the `uri` column,
  - `$1` is `id`,
  - `$2` is `word`
  - and so on

results = conn.gsql('''
  USE GRAPH WordNet
  BEGIN
  CREATE LOADING JOB load_job_WN FOR GRAPH WordNet {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO EDGE type VALUES($0, $1, $2) USING SEPARATOR=",", HEADER="true", EOL="", QUOTE="double";
    }
  END
  ''')
print(results)

#### Word Edges

results = conn.gsql('''
  USE GRAPH WordNet
  BEGIN
  CREATE LOADING JOB load_job_WN_nodes FOR GRAPH WordNet {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX words VALUES($1, $0, $2, $3, $4) USING SEPARATOR=",", HEADER="true", EOL="", QUOTE="double";
    }
  END
  ''')
print(results)

### Load Data

#### Words

# Load the posts file wiht the 'load_words' job
load_words = root+'WN-nodes.csv'
results = conn.uploadFile(load_words, fileTag='MyDataSource', jobName='load_job_WN_nodes')
print(json.dumps(results, indent=2))

#### Edges

# Load the posts file wiht the 'load_edges' job
load_edges = root+'WN-edges.csv'
results = conn.uploadFile(load_edges, fileTag='MyDataSource', jobName='load_job_WN')
print(json.dumps(results, indent=2))

### Exploring the Graph

#### Get Vertex and Edge Schema

results = conn.getVertexTypes()
print(f"Verticies: {results}")
vertices = results

results = conn.getEdgeTypes()
print(f"Edges: {results}")
edges = results


print(f"Results for Post vertex")
pprint(conn.getVertexType("words"))

print("-----------------")
print(f"Results for liked edge")
pprint(conn.getEdgeType("type"))


### Counting Data

print("Vertex Counts")
for vertex in vertices:
  print(f"There are {conn.getVertexCount(vertex)} {vertex} vertices in the graph")

print("--------------")
print("Edge Counts")
for edge in edges:
  print(f"There are {conn.getEdgeCount(edge)} {edge} edges in the graph")

### Extracting Data

#### Vertex/Edge Set Format

##### Getting a Vertex

results = conn.getVerticesById("words", "15314760-n")
pprint(results)

##### Or Multiple Vertices

tdf1 = conn.getVerticesById("words", ["15137796-n","15192825-n"])
pprint(tdf1)

##### Count Edges Connected to a Vertex

results = conn.getEdgeCountFrom("words", "15192825-n")
pprint(results)

##### Show all Edges Connected to a Vertex

results = conn.getEdges("words", "15192825-n")
pprint(results)

#### As Pandas Dataframe
Supports all of the above in native Pandas Dataframe format.

##### All Vertices of one Type

df1 = conn.getVertexDataframe("words")
print(df1)

##### One or More Vertex

df2 = conn.getVertexDataframeById("words", ["15192825-n"])
print(df2)

##### Convert Vertex/Edge Set to Dataframe
We'll use the results from the 'Or Multiple Vertices' cell. 

df3 = conn.vertexSetToDataFrame(tdf1)
print(df3)

##### Get Edges

df4 = conn.getEdgesDataframe("words", "15192825-n", limit=3)
print(df4)

### Path Finding
Find paths between vertices.

Supported are:
- shortestPath - one shortest path between vertices
- allPaths - all paths within the specified edge limit

results = conn.shortestPath([("words", "15192825-n")], [("words", "15161294-n")])
pprint(results)

## ConceptNET

### Create WordNet Graph

results = conn.gsql('CREATE GRAPH ConceptNET(word, synset, is_a, reverse_is_a)')
print(results)

conn.graphname="ConceptNET"
secret = conn.createSecret()
print(secret)
authToken = conn.getToken(secret)
authToken = authToken[0]
print(authToken)
conn = tg.TigerGraphConnection(host=hostName, graphname="ConceptNET", username=userName, password=password, apiToken=authToken)

def pprint(string):
  print(json.dumps(string, indent=2))

#### Create Loading Jobs

#### Concept/ Words Edges

Let's take a look at what one of our files looks like so we can write a loading job. 

- has_synset(FROM rootWord, TO synset)
is_a_synset_of(FROM synset, TO rootWord)


Here it's important to note that the `$0`, `$1` values line up with the columns of your data.
In this example:
- `$0` is the `uri` column,
- `$1` is `id`,
- `$2` is `word`
- and so on

results = conn.gsql('''
  USE GRAPH ConceptNET
  BEGIN
  CREATE LOADING JOB load_job_words FOR GRAPH ConceptNET {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX word VALUES($0, $1, $2) USING SEPARATOR=",", HEADER="true", EOL="", QUOTE="double";
    }

    CREATE LOADING JOB load_job_synsets FOR GRAPH ConceptNET {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX synset VALUES($0, $2) USING SEPARATOR=",", HEADER="true", EOL="", QUOTE="double";
    }


    CREATE LOADING JOB load_job_relationships FOR GRAPH ConceptNET {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO EDGE is_a VALUES($0 word, $1 synset, $4) USING SEPARATOR=",", HEADER="true", EOL="";
      LOAD MyDataSource TO EDGE is_a VALUES($0 synset, $1 word, $4) USING SEPARATOR=",", HEADER="true", EOL="";
    }

  END
  ''')
print(results)

#### Concept/ Word Root


results = conn.gsql('''
  USE GRAPH ConceptNET
  BEGIN
  CREATE LOADING JOB load_job_words_csv_1644515929297 FOR GRAPH ConceptNET {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX word VALUES($0, $1, $2) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
    }

  CREATE LOADING JOB load_job_relationships_csv_1644515956497 FOR GRAPH ConceptNET {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO EDGE is_a VALUES($0 synset, $1 word, $4) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
      LOAD MyDataSource TO EDGE is_a VALUES($0 word, $1 synset, $4) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
    }

  CREATE LOADING JOB load_job_synsets_csv_1644515968121 FOR GRAPH ConceptNET {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO VERTEX synset VALUES($0, $2) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
    }
  END
  ''')
print(results)

#### Concept/ Synset 
- synset(PRIMARY_ID id STRING, pos STRING, definition STRING, label STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"


results = conn.gsql('''
  USE GRAPH ConceptNET
  BEGIN
  CREATE LOADING JOB load_job_relationships FOR GRAPH ConceptNET {
      DEFINE FILENAME MyDataSource;
      LOAD MyDataSource TO EDGE is_a VALUES($0 word, $1 synset, $4) USING SEPARATOR=",", HEADER="true", EOL="\n";
      LOAD MyDataSource TO EDGE is_a VALUES($0 synset, $1 word, $4) USING SEPARATOR=",", HEADER="true", EOL="\n";
    }
  END
  ''')
print(results)

### Load Data

#### Words

# Load the posts file wiht the 'load_words' job
load_words = root+'words.csv'
results = conn.uploadFile(load_words, timeout='100000',fileTag='MyDataSource', jobName='load_job_words_csv_1644515929297')
print(json.dumps(results, indent=2))

#### Synsets

# Load the posts file wiht the 'load_synsets' job
load_synsets = root+'synsets.csv'
results = conn.uploadFile(load_synsets, timeout='100000', fileTag='MyDataSource', jobName='load_job_synsets_csv_1644515968121')
print(json.dumps(results, indent=2))

#### Edges

# Load the posts file wiht the 'load_edges' job
load_edges = root+'relationships.csv'
results = conn.uploadFile(load_edges, timeout='100000', fileTag='MyDataSource', jobName='load_job_relationships_csv_1644515956497')
print(json.dumps(results, indent=2))

### Exploring the Graph

#### Get Vertex and Edge Schema

results = conn.getVertexTypes()
print(f"Verticies: {results}")
vertices = results

results = conn.getEdgeTypes()
print(f"Edges: {results}")
edges = results


# print(f"Results for Post vertex")
# pprint(conn.getVertexType("word"))

print("-----------------")
print(f"Results for liked edge")
pprint(conn.getEdgeType("is_a"))


### Counting Data

print("Vertex Counts")
for vertex in vertices:
  print(f"There are {conn.getVertexCount(vertex)} {vertex} vertices in the graph")

print("--------------")
print("Edge Counts")
for edge in edges:
  print(f"There are {conn.getEdgeCount(edge)} {edge} edges in the graph")

### Extracting Data

#### Vertex/Edge Set Format

##### Getting a Vertex

results = conn.getVerticesById("synset", "judicially.r.01")
pprint(results)

##### Or Multiple Vertices

tdf1 = conn.getVerticesById("word", ["conflict.v","clash.v.02"])
pprint(tdf1)

##### Count Edges Connected to a Vertex

results = conn.getEdgeCountFrom("word", "conflict.v")
pprint(results)

##### Show all Edges Connected to a Vertex

results = conn.getEdges("word", "conflict.v")
pprint(results)

#### As Pandas Dataframe
Supports all of the above in native Pandas Dataframe format.

##### All Vertices of one Type

df1 = conn.getVertexDataframe("word")
print(df1)

##### One or More Vertex

df2 = conn.getVertexDataframeById("word", "conflict.v")
print(df2)

##### Convert Vertex/Edge Set to Dataframe
We'll use the results from the 'Or Multiple Vertices' cell. 

df3 = conn.vertexSetToDataFrame(tdf1)
print(df3)

##### Get Edges

df4 = conn.getEdgesDataframe("word", "conflict.v",limit=3)
print(df4)

### Path Finding
Find paths between vertices.

Supported are:
- shortestPath - one shortest path between vertices
- allPaths - all paths within the specified edge limit

results = conn.shortestPath([("word", "in_due_time.r")], [("word", "in_due_season.r")])
pprint(results)

# Clear the Whole Graph
DANGER ZONE

# conn.gsql('''
# USE GLOBAL
# DROP ALL
# ''')
