==> 6_LanguageModel_CN_UniqueEdge.ipynb <==
# Script to load conceptnet data with unique edges.

*Issue*: loading edges fails for all unique edges. check methods to resolve this.

## Peek into the data 

### ConceptNet - Word

import pandas as pd 

root = 'csv_imports/'

nodes = pd.read_csv(root+'WN-nodes.csv')

nodes.head(1)

### ConceptNet - Synset

synsets = pd.read_csv(root+'synsets.csv')
synsets.head(1)

synsets_enc = pd.read_csv(root+'encoded/synsets-encoded.csv',index_col=0)
synsets_enc.head(1)

#helper method to encode synsets relationship columns.
# def get_synset(id):
#     try:
#       s = id.split('.')
#       return s[0]
#     except:
#       return ""
        
# synsets_enc['name'] = synsets_enc.apply(lambda x: get_synset(x['id:ID']),axis=1)
# synsets_enc.to_csv('encoded/synsets-encoded.csv')

### ConceptNet - Edges

relationships = pd.read_csv(root+'relationships.csv',index_col=[0])
relationships.head(5)

### ConceptNet - Part of Speech

relationships = pd.read_csv(root+'encoded/pos_wn.csv',index_col=[0])
relationships.head()

## Install pyTigerGraph

!pip install -U pyTigerGraph

## Add Imports and Establish Initial Connection

# Imports
import pyTigerGraph as tg
import json
import pandas as pd

# Connection parameters
hostName = "https://language.i.tgcloud.io/"
userName = "tigergraph"
password = "tigergraph"

conn = tg.TigerGraphConnection(host=hostName, username=userName, password=password)

print("Connected")

##Clear graph if already solution has the same graph

# conn.gsql('''
# USE GLOBAL
# DROP ALL
# ''')

## Define and Publish the Schema

- Include the edges for the relation types included: **['Antonym','Synonym','RelatedTo']** to generate the gsql create edge query. 
- This will eliminate the effort to type in queries if there are too many relation types.

rel_type = pd.read_csv(root+'encoded/type_rel.csv')
rel = rel_type['type']
print(len(rel))

rel_included = rel #['Antonym','Synonym','RelatedTo']

str_=""
# for r in rel:
for r in rel_included:
  if '/' in r:
    r = r.replace('/','_')
  str_+="CREATE DIRECTED EDGE "+r+"(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE=\"reverse_"+r+"\""

  str_+="\n"
print(str_)

- Copy & paste the above query within USE Global gsql block after EDGE:is_pos.

# DEFINE / CREATE ALL EDGES AND VERTICES 
results = conn.gsql('''
USE GLOBAL
CREATE VERTEX synset(PRIMARY_ID id STRING, definition STRING, pos STRING,name STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
CREATE VERTEX word(PRIMARY_ID id STRING, name STRING, pos STRING, conceptUri STRING, label STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
CREATE VERTEX part_of_speech(PRIMARY_ID id STRING, type STRING) WITH STATS="OUTDEGREE_BY_EDGETYPE", PRIMARY_ID_AS_ATTRIBUTE="true"
CREATE DIRECTED EDGE has_pos(FROM synset, TO part_of_speech|FROM word, TO part_of_speech) WITH REVERSE_EDGE="reverse_has_pos"
CREATE DIRECTED EDGE is_pos_of(FROM part_of_speech, TO synset|FROM part_of_speech, TO word) WITH REVERSE_EDGE="reverse_is_pos_of"


CREATE DIRECTED EDGE Attribute(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Attribute"
CREATE DIRECTED EDGE InSynset(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_InSynset"
CREATE DIRECTED EDGE Domain(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Domain"
CREATE DIRECTED EDGE SimilarTo(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_SimilarTo"
CREATE DIRECTED EDGE AlsoSee(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_AlsoSee"
CREATE DIRECTED EDGE IsA(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_IsA"
CREATE DIRECTED EDGE PartOf(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_PartOf"
CREATE DIRECTED EDGE Entailment(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Entailment"
CREATE DIRECTED EDGE VerbGroup(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_VerbGroup"
CREATE DIRECTED EDGE Cause(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Cause"
CREATE DIRECTED EDGE Antonym(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Antonym"
CREATE DIRECTED EDGE AtLocation(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_AtLocation"
CREATE DIRECTED EDGE CapableOf(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_CapableOf"
CREATE DIRECTED EDGE Causes(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Causes"
CREATE DIRECTED EDGE CausesDesire(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_CausesDesire"
CREATE DIRECTED EDGE CreatedBy(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_CreatedBy"
CREATE DIRECTED EDGE DefinedAs(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_DefinedAs"
CREATE DIRECTED EDGE DerivedFrom(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_DerivedFrom"
CREATE DIRECTED EDGE Desires(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Desires"
CREATE DIRECTED EDGE DistinctFrom(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_DistinctFrom"
CREATE DIRECTED EDGE Entails(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Entails"
CREATE DIRECTED EDGE EtymologicallyDerivedFrom(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_EtymologicallyDerivedFrom"
CREATE DIRECTED EDGE EtymologicallyRelatedTo(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_EtymologicallyRelatedTo"
CREATE DIRECTED EDGE FormOf(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_FormOf"
CREATE DIRECTED EDGE HasA(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_HasA"
CREATE DIRECTED EDGE HasContext(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_HasContext"
CREATE DIRECTED EDGE HasFirstSubevent(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_HasFirstSubevent"
CREATE DIRECTED EDGE HasLastSubevent(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_HasLastSubevent"
CREATE DIRECTED EDGE HasPrerequisite(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_HasPrerequisite"
CREATE DIRECTED EDGE HasProperty(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_HasProperty"
CREATE DIRECTED EDGE HasSubevent(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_HasSubevent"
CREATE DIRECTED EDGE InstanceOf(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_InstanceOf"
CREATE DIRECTED EDGE LocatedNear(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_LocatedNear"
CREATE DIRECTED EDGE MadeOf(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_MadeOf"
CREATE DIRECTED EDGE MannerOf(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_MannerOf"
CREATE DIRECTED EDGE MotivatedByGoal(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_MotivatedByGoal"
CREATE DIRECTED EDGE NotCapableOf(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_NotCapableOf"
CREATE DIRECTED EDGE NotDesires(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_NotDesires"
CREATE DIRECTED EDGE NotHasProperty(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_NotHasProperty"
CREATE DIRECTED EDGE ReceivesAction(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_ReceivesAction"
CREATE DIRECTED EDGE RelatedTo(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_RelatedTo"
CREATE DIRECTED EDGE SymbolOf(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_SymbolOf"
CREATE DIRECTED EDGE Synonym(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_Synonym"
CREATE DIRECTED EDGE UsedFor(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_UsedFor"
CREATE DIRECTED EDGE dbpedia_capital(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_capital"
CREATE DIRECTED EDGE dbpedia_field(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_field"
CREATE DIRECTED EDGE dbpedia_genre(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_genre"
CREATE DIRECTED EDGE dbpedia_genus(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_genus"
CREATE DIRECTED EDGE dbpedia_influencedBy(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_influencedBy"
CREATE DIRECTED EDGE dbpedia_knownFor(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_knownFor"
CREATE DIRECTED EDGE dbpedia_language(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_language"
CREATE DIRECTED EDGE dbpedia_leader(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_leader"
CREATE DIRECTED EDGE dbpedia_occupation(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_occupation"
CREATE DIRECTED EDGE dbpedia_product(FROM synset, TO synset|FROM synset, TO word|FROM word, TO synset|FROM word, TO word) WITH REVERSE_EDGE="reverse_dbpedia_product"

  ''')
print(results)

## ConceptNet

## Create ConceptNet Graph

- generate the edges for the relation types included: **['Antonym','Synonym','RelatedTo']** to be included in the graph.

rel = rel_type['type']
print(len(rel))
str_=""
# for r in rel:
for r in rel_included:
  if '/' in r:
    r=r.replace('/','_')
  str_+=r+"\n"
print(str_.strip('\n'))

- Copy & paste the above after EDGE: is_pos_of.

results = conn.gsql('CREATE GRAPH ConceptNet(synset, word, part_of_speech,has_pos,is_pos_of,Attribute,InSynset,Domain,SimilarTo,AlsoSee,IsA,PartOf,Entailment,VerbGroup,Cause,Antonym,AtLocation,CapableOf,Causes,CausesDesire,CreatedBy,DefinedAs,DerivedFrom,Desires,DistinctFrom,Entails,EtymologicallyDerivedFrom,EtymologicallyRelatedTo,FormOf,HasA,HasContext,HasFirstSubevent,HasLastSubevent,HasPrerequisite,HasProperty,HasSubevent,InstanceOf,LocatedNear,MadeOf,MannerOf,MotivatedByGoal,NotCapableOf,NotDesires,NotHasProperty,ReceivesAction,RelatedTo,SymbolOf,Synonym,UsedFor,dbpedia_capital,dbpedia_field,dbpedia_genre,dbpedia_genus,dbpedia_influencedBy,dbpedia_knownFor,dbpedia_language,dbpedia_leader,dbpedia_occupation,dbpedia_product)')
print(results)

- Generate AuthToken to update the graph

conn.graphname="ConceptNet"
secret = conn.createSecret()
print(secret)
authToken = conn.getToken(secret)
authToken = authToken[0]
print(authToken)
# authToken = 'rc7reopbis1667ksgcppq5v5fb99p6s1'
conn = tg.TigerGraphConnection(host=hostName, graphname="ConceptNet", username=userName, password=password, apiToken=authToken)

def pprint(string):
  print(json.dumps(string, indent=2))

## Create Loading Jobs

- Generate edge queries for the relation types included: **['Antonym','Synonym','RelatedTo']** to include in the load job gsql.

- Here it's important to note that the `$0`, `$1` values line up with the columns of your data.
In this example:
  - `$0` is the `uri` column,
  - `$1` is `id`,
  - `$2` is `word`
  - and so on

rel_type = pd.read_csv(root+'encoded/type_rel.csv')
rel = rel_type['type']


str_=""
# for r in rel:
for r in rel_included:
  orig=r
  if '/' in r:
    r = r.replace('/','_')
  str_+="LOAD MyDataSource TO EDGE  "+r+" VALUES($0 synset, $1 synset) WHERE $4 == \""+orig+"\" USING SEPARATOR=\",\""", HEADER=\"true\", EOL=""\"\\n"""
  str_+="\";\n"
  str_+="LOAD MyDataSource TO EDGE  "+r+" VALUES($0 synset, $1 word) WHERE $4 == \""+orig+"\" USING SEPARATOR=\",\""", HEADER=\"true\", EOL=""\"\\n"""
  str_+="\";\n"
  str_+="LOAD MyDataSource TO EDGE  "+r+" VALUES($0 word, $1 synset) WHERE $4 == \""+orig+"\" USING SEPARATOR=\",\""", HEADER=\"true\", EOL=""\"\\n"""
  str_+="\";\n"
  str_+="LOAD MyDataSource TO EDGE  "+r+" VALUES($0 word, $1 word) WHERE $4 == \""+orig+"\" USING SEPARATOR=\",\""", HEADER=\"true\", EOL=""\"\\n"""
  str_+="\";\n"
print(str_)

- Copy & Paste the above inside load_job_relationships gsql block.

results = conn.gsql('''
  USE GRAPH ConceptNet
  BEGIN

  CREATE LOADING JOB load_job_relationships FOR GRAPH ConceptNet {
        DEFINE FILENAME MyDataSource;

LOAD MyDataSource TO EDGE  Attribute VALUES($0 synset, $1 synset) WHERE $4 == "Attribute" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Attribute VALUES($0 synset, $1 word) WHERE $4 == "Attribute" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Attribute VALUES($0 word, $1 synset) WHERE $4 == "Attribute" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Attribute VALUES($0 word, $1 word) WHERE $4 == "Attribute" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  InSynset VALUES($0 synset, $1 synset) WHERE $4 == "InSynset" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  InSynset VALUES($0 synset, $1 word) WHERE $4 == "InSynset" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  InSynset VALUES($0 word, $1 synset) WHERE $4 == "InSynset" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  InSynset VALUES($0 word, $1 word) WHERE $4 == "InSynset" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Domain VALUES($0 synset, $1 synset) WHERE $4 == "Domain" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Domain VALUES($0 synset, $1 word) WHERE $4 == "Domain" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Domain VALUES($0 word, $1 synset) WHERE $4 == "Domain" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Domain VALUES($0 word, $1 word) WHERE $4 == "Domain" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  SimilarTo VALUES($0 synset, $1 synset) WHERE $4 == "SimilarTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  SimilarTo VALUES($0 synset, $1 word) WHERE $4 == "SimilarTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  SimilarTo VALUES($0 word, $1 synset) WHERE $4 == "SimilarTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  SimilarTo VALUES($0 word, $1 word) WHERE $4 == "SimilarTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  AlsoSee VALUES($0 synset, $1 synset) WHERE $4 == "AlsoSee" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  AlsoSee VALUES($0 synset, $1 word) WHERE $4 == "AlsoSee" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  AlsoSee VALUES($0 word, $1 synset) WHERE $4 == "AlsoSee" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  AlsoSee VALUES($0 word, $1 word) WHERE $4 == "AlsoSee" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  IsA VALUES($0 synset, $1 synset) WHERE $4 == "IsA" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  IsA VALUES($0 synset, $1 word) WHERE $4 == "IsA" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  IsA VALUES($0 word, $1 synset) WHERE $4 == "IsA" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  IsA VALUES($0 word, $1 word) WHERE $4 == "IsA" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  PartOf VALUES($0 synset, $1 synset) WHERE $4 == "PartOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  PartOf VALUES($0 synset, $1 word) WHERE $4 == "PartOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  PartOf VALUES($0 word, $1 synset) WHERE $4 == "PartOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  PartOf VALUES($0 word, $1 word) WHERE $4 == "PartOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Entailment VALUES($0 synset, $1 synset) WHERE $4 == "Entailment" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Entailment VALUES($0 synset, $1 word) WHERE $4 == "Entailment" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Entailment VALUES($0 word, $1 synset) WHERE $4 == "Entailment" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Entailment VALUES($0 word, $1 word) WHERE $4 == "Entailment" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  VerbGroup VALUES($0 synset, $1 synset) WHERE $4 == "VerbGroup" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  VerbGroup VALUES($0 synset, $1 word) WHERE $4 == "VerbGroup" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  VerbGroup VALUES($0 word, $1 synset) WHERE $4 == "VerbGroup" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  VerbGroup VALUES($0 word, $1 word) WHERE $4 == "VerbGroup" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Cause VALUES($0 synset, $1 synset) WHERE $4 == "Cause" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Cause VALUES($0 synset, $1 word) WHERE $4 == "Cause" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Cause VALUES($0 word, $1 synset) WHERE $4 == "Cause" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Cause VALUES($0 word, $1 word) WHERE $4 == "Cause" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Antonym VALUES($0 synset, $1 synset) WHERE $4 == "Antonym" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Antonym VALUES($0 synset, $1 word) WHERE $4 == "Antonym" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Antonym VALUES($0 word, $1 synset) WHERE $4 == "Antonym" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Antonym VALUES($0 word, $1 word) WHERE $4 == "Antonym" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  AtLocation VALUES($0 synset, $1 synset) WHERE $4 == "AtLocation" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  AtLocation VALUES($0 synset, $1 word) WHERE $4 == "AtLocation" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  AtLocation VALUES($0 word, $1 synset) WHERE $4 == "AtLocation" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  AtLocation VALUES($0 word, $1 word) WHERE $4 == "AtLocation" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CapableOf VALUES($0 synset, $1 synset) WHERE $4 == "CapableOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CapableOf VALUES($0 synset, $1 word) WHERE $4 == "CapableOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CapableOf VALUES($0 word, $1 synset) WHERE $4 == "CapableOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CapableOf VALUES($0 word, $1 word) WHERE $4 == "CapableOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Causes VALUES($0 synset, $1 synset) WHERE $4 == "Causes" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Causes VALUES($0 synset, $1 word) WHERE $4 == "Causes" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Causes VALUES($0 word, $1 synset) WHERE $4 == "Causes" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Causes VALUES($0 word, $1 word) WHERE $4 == "Causes" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CausesDesire VALUES($0 synset, $1 synset) WHERE $4 == "CausesDesire" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CausesDesire VALUES($0 synset, $1 word) WHERE $4 == "CausesDesire" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CausesDesire VALUES($0 word, $1 synset) WHERE $4 == "CausesDesire" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CausesDesire VALUES($0 word, $1 word) WHERE $4 == "CausesDesire" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CreatedBy VALUES($0 synset, $1 synset) WHERE $4 == "CreatedBy" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CreatedBy VALUES($0 synset, $1 word) WHERE $4 == "CreatedBy" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CreatedBy VALUES($0 word, $1 synset) WHERE $4 == "CreatedBy" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  CreatedBy VALUES($0 word, $1 word) WHERE $4 == "CreatedBy" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DefinedAs VALUES($0 synset, $1 synset) WHERE $4 == "DefinedAs" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DefinedAs VALUES($0 synset, $1 word) WHERE $4 == "DefinedAs" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DefinedAs VALUES($0 word, $1 synset) WHERE $4 == "DefinedAs" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DefinedAs VALUES($0 word, $1 word) WHERE $4 == "DefinedAs" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DerivedFrom VALUES($0 synset, $1 synset) WHERE $4 == "DerivedFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DerivedFrom VALUES($0 synset, $1 word) WHERE $4 == "DerivedFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DerivedFrom VALUES($0 word, $1 synset) WHERE $4 == "DerivedFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DerivedFrom VALUES($0 word, $1 word) WHERE $4 == "DerivedFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Desires VALUES($0 synset, $1 synset) WHERE $4 == "Desires" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Desires VALUES($0 synset, $1 word) WHERE $4 == "Desires" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Desires VALUES($0 word, $1 synset) WHERE $4 == "Desires" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Desires VALUES($0 word, $1 word) WHERE $4 == "Desires" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DistinctFrom VALUES($0 synset, $1 synset) WHERE $4 == "DistinctFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DistinctFrom VALUES($0 synset, $1 word) WHERE $4 == "DistinctFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DistinctFrom VALUES($0 word, $1 synset) WHERE $4 == "DistinctFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  DistinctFrom VALUES($0 word, $1 word) WHERE $4 == "DistinctFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Entails VALUES($0 synset, $1 synset) WHERE $4 == "Entails" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Entails VALUES($0 synset, $1 word) WHERE $4 == "Entails" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Entails VALUES($0 word, $1 synset) WHERE $4 == "Entails" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Entails VALUES($0 word, $1 word) WHERE $4 == "Entails" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  EtymologicallyDerivedFrom VALUES($0 synset, $1 synset) WHERE $4 == "EtymologicallyDerivedFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  EtymologicallyDerivedFrom VALUES($0 synset, $1 word) WHERE $4 == "EtymologicallyDerivedFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  EtymologicallyDerivedFrom VALUES($0 word, $1 synset) WHERE $4 == "EtymologicallyDerivedFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  EtymologicallyDerivedFrom VALUES($0 word, $1 word) WHERE $4 == "EtymologicallyDerivedFrom" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  EtymologicallyRelatedTo VALUES($0 synset, $1 synset) WHERE $4 == "EtymologicallyRelatedTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  EtymologicallyRelatedTo VALUES($0 synset, $1 word) WHERE $4 == "EtymologicallyRelatedTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  EtymologicallyRelatedTo VALUES($0 word, $1 synset) WHERE $4 == "EtymologicallyRelatedTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  EtymologicallyRelatedTo VALUES($0 word, $1 word) WHERE $4 == "EtymologicallyRelatedTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  FormOf VALUES($0 synset, $1 synset) WHERE $4 == "FormOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  FormOf VALUES($0 synset, $1 word) WHERE $4 == "FormOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  FormOf VALUES($0 word, $1 synset) WHERE $4 == "FormOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  FormOf VALUES($0 word, $1 word) WHERE $4 == "FormOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasA VALUES($0 synset, $1 synset) WHERE $4 == "HasA" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasA VALUES($0 synset, $1 word) WHERE $4 == "HasA" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasA VALUES($0 word, $1 synset) WHERE $4 == "HasA" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasA VALUES($0 word, $1 word) WHERE $4 == "HasA" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasContext VALUES($0 synset, $1 synset) WHERE $4 == "HasContext" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasContext VALUES($0 synset, $1 word) WHERE $4 == "HasContext" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasContext VALUES($0 word, $1 synset) WHERE $4 == "HasContext" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasContext VALUES($0 word, $1 word) WHERE $4 == "HasContext" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasFirstSubevent VALUES($0 synset, $1 synset) WHERE $4 == "HasFirstSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasFirstSubevent VALUES($0 synset, $1 word) WHERE $4 == "HasFirstSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasFirstSubevent VALUES($0 word, $1 synset) WHERE $4 == "HasFirstSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasFirstSubevent VALUES($0 word, $1 word) WHERE $4 == "HasFirstSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasLastSubevent VALUES($0 synset, $1 synset) WHERE $4 == "HasLastSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasLastSubevent VALUES($0 synset, $1 word) WHERE $4 == "HasLastSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasLastSubevent VALUES($0 word, $1 synset) WHERE $4 == "HasLastSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasLastSubevent VALUES($0 word, $1 word) WHERE $4 == "HasLastSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasPrerequisite VALUES($0 synset, $1 synset) WHERE $4 == "HasPrerequisite" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasPrerequisite VALUES($0 synset, $1 word) WHERE $4 == "HasPrerequisite" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasPrerequisite VALUES($0 word, $1 synset) WHERE $4 == "HasPrerequisite" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasPrerequisite VALUES($0 word, $1 word) WHERE $4 == "HasPrerequisite" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasProperty VALUES($0 synset, $1 synset) WHERE $4 == "HasProperty" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasProperty VALUES($0 synset, $1 word) WHERE $4 == "HasProperty" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasProperty VALUES($0 word, $1 synset) WHERE $4 == "HasProperty" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasProperty VALUES($0 word, $1 word) WHERE $4 == "HasProperty" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasSubevent VALUES($0 synset, $1 synset) WHERE $4 == "HasSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasSubevent VALUES($0 synset, $1 word) WHERE $4 == "HasSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasSubevent VALUES($0 word, $1 synset) WHERE $4 == "HasSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  HasSubevent VALUES($0 word, $1 word) WHERE $4 == "HasSubevent" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  InstanceOf VALUES($0 synset, $1 synset) WHERE $4 == "InstanceOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  InstanceOf VALUES($0 synset, $1 word) WHERE $4 == "InstanceOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  InstanceOf VALUES($0 word, $1 synset) WHERE $4 == "InstanceOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  InstanceOf VALUES($0 word, $1 word) WHERE $4 == "InstanceOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  LocatedNear VALUES($0 synset, $1 synset) WHERE $4 == "LocatedNear" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  LocatedNear VALUES($0 synset, $1 word) WHERE $4 == "LocatedNear" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  LocatedNear VALUES($0 word, $1 synset) WHERE $4 == "LocatedNear" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  LocatedNear VALUES($0 word, $1 word) WHERE $4 == "LocatedNear" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MadeOf VALUES($0 synset, $1 synset) WHERE $4 == "MadeOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MadeOf VALUES($0 synset, $1 word) WHERE $4 == "MadeOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MadeOf VALUES($0 word, $1 synset) WHERE $4 == "MadeOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MadeOf VALUES($0 word, $1 word) WHERE $4 == "MadeOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MannerOf VALUES($0 synset, $1 synset) WHERE $4 == "MannerOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MannerOf VALUES($0 synset, $1 word) WHERE $4 == "MannerOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MannerOf VALUES($0 word, $1 synset) WHERE $4 == "MannerOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MannerOf VALUES($0 word, $1 word) WHERE $4 == "MannerOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MotivatedByGoal VALUES($0 synset, $1 synset) WHERE $4 == "MotivatedByGoal" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MotivatedByGoal VALUES($0 synset, $1 word) WHERE $4 == "MotivatedByGoal" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MotivatedByGoal VALUES($0 word, $1 synset) WHERE $4 == "MotivatedByGoal" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  MotivatedByGoal VALUES($0 word, $1 word) WHERE $4 == "MotivatedByGoal" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotCapableOf VALUES($0 synset, $1 synset) WHERE $4 == "NotCapableOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotCapableOf VALUES($0 synset, $1 word) WHERE $4 == "NotCapableOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotCapableOf VALUES($0 word, $1 synset) WHERE $4 == "NotCapableOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotCapableOf VALUES($0 word, $1 word) WHERE $4 == "NotCapableOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotDesires VALUES($0 synset, $1 synset) WHERE $4 == "NotDesires" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotDesires VALUES($0 synset, $1 word) WHERE $4 == "NotDesires" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotDesires VALUES($0 word, $1 synset) WHERE $4 == "NotDesires" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotDesires VALUES($0 word, $1 word) WHERE $4 == "NotDesires" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotHasProperty VALUES($0 synset, $1 synset) WHERE $4 == "NotHasProperty" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotHasProperty VALUES($0 synset, $1 word) WHERE $4 == "NotHasProperty" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotHasProperty VALUES($0 word, $1 synset) WHERE $4 == "NotHasProperty" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  NotHasProperty VALUES($0 word, $1 word) WHERE $4 == "NotHasProperty" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  ReceivesAction VALUES($0 synset, $1 synset) WHERE $4 == "ReceivesAction" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  ReceivesAction VALUES($0 synset, $1 word) WHERE $4 == "ReceivesAction" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  ReceivesAction VALUES($0 word, $1 synset) WHERE $4 == "ReceivesAction" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  ReceivesAction VALUES($0 word, $1 word) WHERE $4 == "ReceivesAction" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  RelatedTo VALUES($0 synset, $1 synset) WHERE $4 == "RelatedTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  RelatedTo VALUES($0 synset, $1 word) WHERE $4 == "RelatedTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  RelatedTo VALUES($0 word, $1 synset) WHERE $4 == "RelatedTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  RelatedTo VALUES($0 word, $1 word) WHERE $4 == "RelatedTo" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  SymbolOf VALUES($0 synset, $1 synset) WHERE $4 == "SymbolOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  SymbolOf VALUES($0 synset, $1 word) WHERE $4 == "SymbolOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  SymbolOf VALUES($0 word, $1 synset) WHERE $4 == "SymbolOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  SymbolOf VALUES($0 word, $1 word) WHERE $4 == "SymbolOf" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Synonym VALUES($0 synset, $1 synset) WHERE $4 == "Synonym" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Synonym VALUES($0 synset, $1 word) WHERE $4 == "Synonym" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Synonym VALUES($0 word, $1 synset) WHERE $4 == "Synonym" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  Synonym VALUES($0 word, $1 word) WHERE $4 == "Synonym" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  UsedFor VALUES($0 synset, $1 synset) WHERE $4 == "UsedFor" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  UsedFor VALUES($0 synset, $1 word) WHERE $4 == "UsedFor" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  UsedFor VALUES($0 word, $1 synset) WHERE $4 == "UsedFor" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  UsedFor VALUES($0 word, $1 word) WHERE $4 == "UsedFor" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_capital VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/capital" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_capital VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/capital" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_capital VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/capital" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_capital VALUES($0 word, $1 word) WHERE $4 == "dbpedia/capital" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_field VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/field" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_field VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/field" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_field VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/field" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_field VALUES($0 word, $1 word) WHERE $4 == "dbpedia/field" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_genre VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/genre" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_genre VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/genre" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_genre VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/genre" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_genre VALUES($0 word, $1 word) WHERE $4 == "dbpedia/genre" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_genus VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/genus" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_genus VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/genus" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_genus VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/genus" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_genus VALUES($0 word, $1 word) WHERE $4 == "dbpedia/genus" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_influencedBy VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/influencedBy" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_influencedBy VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/influencedBy" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_influencedBy VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/influencedBy" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_influencedBy VALUES($0 word, $1 word) WHERE $4 == "dbpedia/influencedBy" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_knownFor VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/knownFor" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_knownFor VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/knownFor" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_knownFor VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/knownFor" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_knownFor VALUES($0 word, $1 word) WHERE $4 == "dbpedia/knownFor" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_language VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/language" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_language VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/language" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_language VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/language" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_language VALUES($0 word, $1 word) WHERE $4 == "dbpedia/language" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_leader VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/leader" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_leader VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/leader" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_leader VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/leader" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_leader VALUES($0 word, $1 word) WHERE $4 == "dbpedia/leader" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_occupation VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/occupation" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_occupation VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/occupation" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_occupation VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/occupation" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_occupation VALUES($0 word, $1 word) WHERE $4 == "dbpedia/occupation" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_product VALUES($0 synset, $1 synset) WHERE $4 == "dbpedia/product" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_product VALUES($0 synset, $1 word) WHERE $4 == "dbpedia/product" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_product VALUES($0 word, $1 synset) WHERE $4 == "dbpedia/product" USING SEPARATOR=",", HEADER="true", EOL="\n";
LOAD MyDataSource TO EDGE  dbpedia_product VALUES($0 word, $1 word) WHERE $4 == "dbpedia/product" USING SEPARATOR=",", HEADER="true", EOL="\n";
      }


  CREATE LOADING JOB load_job_pos_wn FOR GRAPH ConceptNet {
        DEFINE FILENAME MyDataSource;
        LOAD MyDataSource TO VERTEX part_of_speech VALUES($2, $1) USING SEPARATOR=",", HEADER="true", EOL="\n";
      }


  CREATE LOADING JOB load_job_synsets_encoded FOR GRAPH ConceptNet {
        DEFINE FILENAME MyDataSource;
        LOAD MyDataSource TO VERTEX synset VALUES($0, $2, $1, $4) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
      }

  CREATE LOADING JOB load_job_words_encoded FOR GRAPH ConceptNet {
        DEFINE FILENAME MyDataSource;
        LOAD MyDataSource TO VERTEX word VALUES($0, $1, $2, $3, $4) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
        LOAD MyDataSource TO EDGE has_pos VALUES($0 word, $2 part_of_speech) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
        LOAD MyDataSource TO EDGE is_pos_of VALUES($2 part_of_speech, $0 word) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
        LOAD MyDataSource TO EDGE has_pos VALUES($0 synset, $2 part_of_speech) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
        LOAD MyDataSource TO EDGE is_pos_of VALUES($2 part_of_speech, $0 synset) USING SEPARATOR=",", HEADER="true", EOL="\n", QUOTE="double";
      }


  END
  ''')
print(results)

## Load Data

### load relationships

# Load the posts file wiht the 'load_words' job
load_words = root+'relationships.csv'
results = conn.uploadFile(load_words, timeout=1000000,fileTag='MyDataSource', jobName='load_job_relationships')
# print(json.dumps(results, indent=2))

### load pos

# Load the posts file wiht the 'load_words' job
load_words = root+'encoded/pos_wn.csv'
results = conn.uploadFile(load_words, timeout=100000,fileTag='MyDataSource', jobName='load_job_pos_wn')
print(json.dumps(results, indent=2))

### load words and synsets

# Load the posts file wiht the 'load_edges' job
load_edges = root+'encoded/synsets-encoded.csv'
results = conn.uploadFile(load_edges, timeout=100000, fileTag='MyDataSource', jobName='load_job_synsets_encoded')
print(json.dumps(results, indent=2))

# Load the posts file wiht the 'load_edges' job
load_edges = root+'encoded/words_encoded.csv'
results = conn.uploadFile(load_edges, timeout=100000, fileTag='MyDataSource', jobName='load_job_words_encoded')
print(json.dumps(results, indent=2))

## Exploring the Graph

### Get Vertex and Edge Schema

results = conn.getVertexTypes()
print(f"Verticies: {results}")
vertices = results

results = conn.getEdgeTypes()
print(f"Edges: {results}")
edges = results


print(f"Results for Post vertex")
pprint(conn.getVertexType("word"))

print("-----------------")
print(f"Results for liked edge")
pprint(conn.getEdgeType("type"))


## Counting Data

print("Vertex Counts")
for vertex in vertices:
  print(f"There are {conn.getVertexCount(vertex)} {vertex} vertices in the graph")

print("--------------")
print("Edge Counts")
for edge in edges:
  print(f"There are {conn.getEdgeCount(edge)} {edge} edges in the graph")

# Clear the Whole Graph
DANGER ZONE

# conn.gsql('''
# USE GLOBAL
# DROP ALL
# ''')
